added a missing const qualifier, removed trailing whitespace.
2003-09-30  Sven Neumann  <sven@gimp.org>

	* gap/gap_match.c (p_match_number): added a missing const qualifier,
	removed trailing whitespace.

2003-09-28  Wolfgang Hofer <hof@gimp.org>
        MovePath: dialog layout reorganisation
	bugfixes:
	- report errors when controlpointfile load/save failed.
	- overwrite warning dialog if controlpointfile already exists.
	new  features (some are not completely finished yet)
	- Instant Apply
	- Perspective Transformation
	- Stepspeed Factor
	- Tween Layer Processing for fast moving objects
	  ==> Rendering not implemented yet
	- TraceLayers
	  ==> Rendering not implemented yet
	- changed Scalefactors, Opacity and Rotation values
	  from int to float (int precentage 0 to 100 was not precise enough)
	- extended the controlpoint fileformat to handle float numbers
	  and perspective factors (old fileformat is also accepted at load)

 	* gap/gap_main.c
 	* gap/gap_mov_dialog.c [.h]
 	* gap/gap_mov_exec.c [.h]
 	* gap/gap_mov_render.c [.h]
 	* gap/gap_arr_dialog.c [.h]

+++ gap/gap_mov_exec.h
@@ -25,6 +25,7 @@
  */
 
 /* revision history:
+ * gimp    1.3.20c; 2003/09/23  hof: changed startangle from int to gdouble
  * gimp    1.3.5a;  2002/04/20  hof: API cleanup
  * gimp    1.1.29b; 2000/11/25  hof: NONINTEACTIV PDB interface for Movepath
  * gimp    1.1.20a; 2000/04/25  hof: support for keyframes, anim_preview
@@ -40,7 +41,7 @@
 #include "libgimp/gimp.h"
 #include "gap_mov_dialog.h"
 
-gint32  gap_move_path(GimpRunMode run_mode, gint32 image_id, t_mov_values *pvals, gchar *pointfile, gint rotation_follow, gint32 startangle);
+gint32  gap_move_path(GimpRunMode run_mode, gint32 image_id, t_mov_values *pvals, gchar *pointfile, gint rotation_follow, gdouble startangle);
 gint32  p_mov_anim_preview(t_mov_values *pvals_orig, t_anim_info *ainfo_ptr, gint preview_frame_nr);
 
 gchar  *p_gap_chk_keyframes(t_mov_values *pvals);
@@ -48,7 +49,7 @@ gint    p_conv_keyframe_to_rel(gint abs_keyframe, t_mov_values *pvals);
 gint    p_conv_keyframe_to_abs(gint rel_keyframe, t_mov_values *pvals);
 gint    p_gap_save_pointfile(char *filename, t_mov_values *pvals);
 gint    p_gap_load_pointfile(char *filename, t_mov_values *pvals);
-void    p_calculate_rotate_follow(t_mov_values *pvals, gint32 startangle);
+void    p_calculate_rotate_follow(t_mov_values *pvals, gdouble startangle);
 
 
 #endif
